add MipMapGenerator helper class and caching for device
update Texture generateMipMaps to use new cached device generator
